refactor: Move modules around
